Enum console_api::field::Value

pub enum Value {
    DebugVal(String),
    StrVal(String),
    U64Val(u64),
    I64Val(i64),
    BoolVal(bool),
}

The value of the key-value pair.

Variants

DebugVal(String)

A value serialized to a string using fmt::Debug.

StrVal(String)

A string value.

U64Val(u64)

An unsigned integer value.

I64Val(i64)

A signed integer value.

BoolVal(bool)

A boolean value.

Implementations

impl Value

pub fn encode<B>(&self, buf: &mut B)
where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<Value>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>
where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.