Module console_api::instrument

Represents interactions between the console-subscriber and a console client observing it.

Modules

instrument_client

Generated client implementations.

instrument_server

Generated server implementations.

Structs

InstrumentRequest

InstrumentRequest requests the stream of updates to observe the async runtime state over time.

PauseRequest

PauseRequest requests the stream of updates to pause.

PauseResponse

PauseResponse is the value returned after a pause request.

ResumeRequest

ResumeRequest requests the stream of updates to resume after a pause.

ResumeResponse

ResumeResponse is the value returned after a resume request.

TaskDetailsRequest

TaskDetailsRequest requests the stream of updates about the specific task identified in the request.

Update

Update carries all information regarding tasks, resources, async operations and resource operations in one message. There are a couple of reasons to combine all of these into a single message: