Enum console_api::tasks::task::Kind

#[repr(i32)]
pub enum Kind {
    Spawn = 0,
    Blocking = 1,
}

The category of task this task belongs to.

Variants

Spawn = 0

A task spawned using a runtime’s standard asynchronous task spawning operation (such as tokio::task::spawn).

Blocking = 1

A task spawned via a runtime’s blocking task spawning operation (such as tokio::task::spawn_blocking).

Implementations

impl Kind

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Kind.

pub fn from_i32(value: i32) -> Option<Kind>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Kind, or None if value is not a valid variant.

impl Kind

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.