Trait console_api::instrument::instrument_server::Instrument

pub trait Instrument: Send + Sync + 'static {
    type WatchUpdatesStream: Stream<Item = Result<Update, Status>> + Send + 'static;
    type WatchTaskDetailsStream: Stream<Item = Result<TaskDetails, Status>> + Send + 'static;

    // Required methods
    fn watch_updates<'life0, 'async_trait>(
        &'life0 self,
        request: Request<InstrumentRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::WatchUpdatesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn watch_task_details<'life0, 'async_trait>(
        &'life0 self,
        request: Request<TaskDetailsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::WatchTaskDetailsStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pause<'life0, 'async_trait>(
        &'life0 self,
        request: Request<PauseRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PauseResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn resume<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ResumeRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ResumeResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with InstrumentServer.

Required Associated Types

type WatchUpdatesStream: Stream<Item = Result<Update, Status>> + Send + 'static

Server streaming response type for the WatchUpdates method.

type WatchTaskDetailsStream: Stream<Item = Result<TaskDetails, Status>> + Send + 'static

Server streaming response type for the WatchTaskDetails method.

Required Methods

fn watch_updates<'life0, 'async_trait>( &'life0 self, request: Request<InstrumentRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<Self::WatchUpdatesStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Produces a stream of updates representing the behavior of the instrumented async runtime.

fn watch_task_details<'life0, 'async_trait>( &'life0 self, request: Request<TaskDetailsRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<Self::WatchTaskDetailsStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Produces a stream of updates describing the activity of a specific task.

fn pause<'life0, 'async_trait>( &'life0 self, request: Request<PauseRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<PauseResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Registers that the console observer wants to pause the stream.

fn resume<'life0, 'async_trait>( &'life0 self, request: Request<ResumeRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ResumeResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Registers that the console observer wants to resume the stream.